Updated the handling of MX-records.
There was a bug in the handling of MX-records which had been
present right from the start - because I was following the
TinyDNS spec, rather than what I intended to write.

The difference is that we should parse this:

        @name:destination:priority:TTL

Instead we'd been parsing:

        @name.tld::destination:priority

Note the second field is empty, and we would ignore any optional
TTL-record?

Updated to be correct as per the documentation, and the test-cases
now validate that.